+++ attributes/default.rb
@@ -48,7 +48,9 @@ default['rs-storage']['device']['volume_type'] = nil
 default['rs-storage']['device']['controller_type'] = nil
 
 # The filesystem to be used on the device
-default['rs-storage']['device']['filesystem'] = 'ext4'
+# RHEL 7 and CentOS 7 uses XFS as their default file system.
+default['rs-storage']['device']['filesystem'] =
+  node['platform_family'] == 'rhel' && node['platform_version'] =~ /^7\./ ? 'xfs' : 'ext4'
 
 # Amount of time (in seconds) to wait for a volume to detach at decommission
 default['rs-storage']['device']['detach_timeout'] = 300
@@ -56,9 +58,10 @@ default['rs-storage']['device']['detach_timeout'] = 300
 # Whether to destroy volume(s) on decommission
 default['rs-storage']['device']['destroy_on_decommission'] = false
 
-# The additional options/flags to use for the `mkfs` command. If the whole device is formatted, the force (-F) flag
-# can be used (on ext4 filesystem) to force the operation. This flag may vary based on the filesystem type.
-default['rs-storage']['device']['mkfs_options'] = '-F'
+# The additional options/flags to use for the `mkfs` command. If the whole device is formatted, the force flag
+# can be used to force the operation. This flag varies based on the filesystem type.
+default['rs-storage']['device']['mkfs_options'] =
+  node['rs-storage']['device']['filesystem'] == 'xfs' ? '-f' : '-F'
 
 # The stripe size to use while creating LVM
 default['rs-storage']['device']['stripe_size'] = 512