+++ spec/schedule_spec.rb
@@ -3,7 +3,7 @@ require_relative 'spec_helper.rb'
 describe 'rs-storage::schedule' do
   context 'rs-storage/schedule/enable is true' do
     let(:chef_run) do
-      ChefSpec::Runner.new do |node|
+      ChefSpec::SoloRunner.new do |node|
         node.set['rs-storage']['schedule']['enable'] = true
         node.set['rs-storage']['backup']['lineage'] = 'testing'
         node.set['rs-storage']['schedule']['hour'] = '10'
@@ -16,14 +16,14 @@ describe 'rs-storage::schedule' do
       expect(chef_run).to create_cron("backup_schedule_#{lineage}").with(
         minute: chef_run.node['rs-storage']['schedule']['minute'],
         hour: chef_run.node['rs-storage']['schedule']['hour'],
-        command: "rs_run_recipe --policy 'rs-storage::backup' --name 'rs-storage::backup'"
+        command: "sudo rsc rl10 run_right_script /rll/run/right_script 'right_script=Storage Toolbox Backup - chef'"
       )
     end
   end
 
   context 'rs-storage/schedule/enable is false' do
     let(:chef_run) do
-      ChefSpec::Runner.new do |node|
+      ChefSpec::SoloRunner.new do |node|
         node.set['rs-storage']['schedule']['enable'] = false
         node.set['rs-storage']['backup']['lineage'] = 'testing'
       end.converge(described_recipe)
@@ -32,14 +32,14 @@ describe 'rs-storage::schedule' do
 
     it 'deletes a crontab entry' do
       expect(chef_run).to delete_cron("backup_schedule_#{lineage}").with(
-        command: "rs_run_recipe --policy 'rs-storage::backup' --name 'rs-storage::backup'"
+        command: "sudo rsc rl10 run_right_script /rll/run/right_script 'right_script=Storage Toolbox Backup - chef'"
       )
     end
   end
 
   context 'rs-storage/schedule/hour or rs-storage/schedule/minute missing' do
     let(:chef_run) do
-      ChefSpec::Runner.new do |node|
+      ChefSpec::SoloRunner.new do |node|
         node.set['rs-storage']['backup']['lineage'] = 'testing'
         node.set['rs-storage']['schedule']['enable'] = true
         node.set['rs-storage']['schedule']['hour'] = '10'